Update src/main/java/io/github/ardoco/textproviderjson/converter/ObjectToDtoConverter.java
1L instead of (long) 1

+++ src/main/java/io/github/ardoco/textproviderjson/converter/ObjectToDtoConverter.java
@@ -145,7 +145,7 @@ public class ObjectToDtoConverter {
     private OutgoingDependencyDTO convertToDepOutDTO(DependencyImpl dependency) {
         OutgoingDependencyDTO dependencyDTO = new OutgoingDependencyDTO();
         dependencyDTO.setDependencyTag(dependency.getDependencyTag());
-        dependencyDTO.setTargetWordId(dependency.getWordId() + (long) 1);
+        dependencyDTO.setTargetWordId(dependency.getWordId() + 1L);
         return dependencyDTO;
     }
 }
